pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ovmobilebench"
version = "0.1.0"
description = "End-to-end benchmarking pipeline for OpenVINO on mobile devices"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Embedded Dev Research Team"},
]
dependencies = [
    "typer>=0.9.0",
    "click>=8.1.0",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.2",
    "paramiko>=3.4.0",
    "pandas>=2.2.2",
    "rich>=13.7.1",
    "adbutils>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
    "black>=24.4.2",
    "types-PyYAML>=6.0.12",
]

[project.scripts]
ovmobilebench = "ovmobilebench.cli:app"

[tool.setuptools.packages.find]
include = ["ovmobilebench*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["ovmobilebench", "scripts"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "scripts/setup_ssh_ci.sh",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
]
